+++ youtora/collect/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ScrapeAppConfig(AppConfig):
+    name = "youtora.scrape"
