--- youtora/parse/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ParseConfig(AppConfig):
-    name = 'youtora.parse'
