+++ youtora/refine/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ParseConfig(AppConfig):
+    name = 'youtora.refine'
