+++ youtora/refine/errors.py
@@ -1,0 +1,6 @@
+
+class CaptionNotFoundError(Exception):
+    """
+    an Exception to be raised when the desired caption was not found
+    """
+    pass
