--- youtora/scrape/errors.py
@@ -1,6 +1,0 @@
-
-class CaptionNotFoundError(Exception):
-    """
-    an Exception to be raised when the desired caption was not found
-    """
-    pass
